Sync progress directly to parent's eventProgress
As long as we're keeping track of progress in the EventTasks component,
we may as well use that value in the TaskProgress component too. By
emitting `update:<field>` we can let Vue take care of updating the
correct propery in the parent component while removing the need to
manually call setProgress().

This enables progress data for events to be loaded from the user's local
storage at runtime since the values are now updated with eventProgress.