Minor text fixes
This adds a placeholder when there are no tasks for an event so that
each card can butt up against the previous one to form one long list.

Also increases the height of the progress bar so that it looks aligns
with the task descriptions, and adds some margin to the progress text if
it's being displayed outside of the progress bar itself.