Moar progress
Adds an extra progress bar that shows the total number of completed
tasks out of the number of tasks in any given event, along with the
actual numbers shown in the top right of the card header.

Also tweaks the colouring of the task progress bars and buttons for a
slightly more muted feel, which looks much better than red/blue/green.