Add ability to toggle an input on progress bars
This adds both an inner and outer number input with the value set to the
current progress of that task. We're not using `v-model` here because if
we were to break the 50% mark in any direction then it would switch, and
that's not something we want as it loses focus on the input switched out.