Update events data
Fixes the wording of tasks from notes to reflect the in-game tasks and
adds a whole bunch of tasks that were missing from notes either because
they were new (Weekly Challenge reset?) or I already completed them.

Also reapplies the IDs to make future maintenance easier. We're storing
task progress grouped into objects in local storage based on the event
they're part of, so there's no need for a continuous incrementation as
we can simply reset at 1 instead of finding the last ID that was used.