Replace dashboard view with EventsList component
Another thing that was merely serving as a wrapper and doing nothing
other than pulling in the component. It would've made more sense if
there were other things to put on the page, but since it's only the
events list there's really not much point dedicating a component to it.