Add basic target hints to event progress bar
This adds 'blobs' to the main progress bar for each event positioned in
line with the % of points needed to obtain that particular reward.

The requirement and reward is shown on hover with the title tag for now,
but this would be better off as a proper tooltip in future. Perhaps with
the required number of points visible without hovering at all...

Still need to figure out how to handle the actual progress. The current
bar alone is no good as it tracks not the completed tasks, but *all*
combined task requirements as one huge number. Will either need to
replace the main progress bar for events with targets, or add a second.