Add target tracking progress bar
This updates the `completed` task count to instead hold an array of
completed task objects, which enables us to get a sum of the points
gained from completing tasks in an event that has point-based targets.

With that we add a second progress bar in yellow that will increase by
the number of points gained when any given task is completed. It's by no
means perfect yet as we still need to account for the width of the main
progress bar. Until then, it could end up pushed off the end of the bar.