Fix unnamed event with 0/0 progress and pad body
If an event is added with targets, but the related tasks array has a
single empty object then instead of the "no tasks" message, we get a
task with no description or reward that has progress set to 0/0.

This fixes the issue by checking the length of the keys array for the
first object within the event.tasks array if the given event has one.